rename category actions
